fix: update order of schema exports
I hate you TS compiler.
