+++ packages/schema/src/collection/validation/index.ts
@@ -1,2 +1,2 @@
-export * from "./collection-action-validation"
 export * from "./collection-ut-validation"
+export * from "./collection-action-validation"
