+++ packages/schema/src/room/room-variant.ts
@@ -1,4 +1,4 @@
-import z from "zod"
+import { z } from "zod"
 
 // schemas
 import { roomPlayerSchema } from "@/player"
