+++ packages/schema/src/session/index.ts
@@ -1,2 +1,3 @@
 export * from "./session-schema"
 export * from "./session-variant"
+export * from "./session-metadata"
